css/image-fit/js/ua-stylesheet/video/005.html

<!doctype html>
<title>video default style, controls</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video controls></video>
<script>
test(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'contain');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
